perf(engine): bf16 weights + cuBLAS GEMM + FlashAttention-2
Drop INT8 weight-quant; load all weights bf16 and run the whole activation
pipeline in bf16 (fair bf16-parity with vLLM). Replace the hand-rolled
WMMA/GEMV matmuls and the decode/prefill split with a single cuBLAS
cublasGemmEx path (bf16 in, fp32 accumulate). Replace the per-(head,row)
varlen attention with a FlashAttention-2 style paged kernel: one block per
(q-tile, head, request), BM query rows share BLOCK-sized K/V tiles staged in
shared memory, online softmax + deferred normalization, causal block-skip.

conc=32 (1024/128, temp0): 87.4 -> 350.0 tok/s (4x; 12.5% -> 50.1% of vLLM,
3.3x main). conc=1: 24.9 -> 38.2 (bf16 weights read 2x INT8, so single-stream
trails main's INT8 56.7 — the expected de-quant cost; conc=32 is the target).

+++ engine/src/block_pool.cu
@@ -52,7 +52,7 @@ BlockPool::~BlockPool() {
 // ---- write side: scatter freshly-projected K/V rows into the pool -----------------------
 // The pool's only mutating compute op (the attention kernels read it; see kernels.cu). Token m
 // belongs to the sequence whose block table is row bt_row[m] of `bt`, at logical position pos[m].
-__global__ void store_kv_paged_kernel(const float* __restrict__ src, bf16* __restrict__ cache,
+__global__ void store_kv_paged_kernel(const bf16* __restrict__ src, bf16* __restrict__ cache,
                                       const int* __restrict__ bt, int max_blocks, int block_size,
                                       int kv_dim, const int* __restrict__ bt_row,
                                       const int* __restrict__ pos, int M) {
@@ -60,10 +60,10 @@ __global__ void store_kv_paged_kernel(const float* __restrict__ src, bf16* __res
     if (idx >= M * kv_dim) return;
     int m = idx / kv_dim, i = idx % kv_dim;
     size_t phys = kv_phys_row(bt + (size_t)bt_row[m] * max_blocks, pos[m], block_size);
-    cache[phys * kv_dim + i] = __float2bfloat16(src[idx]);
+    cache[phys * kv_dim + i] = src[idx];
 }
 
-void launch_store_kv_paged(const float* src, bf16* cache, const int* bt, int max_blocks,
+void launch_store_kv_paged(const bf16* src, bf16* cache, const int* bt, int max_blocks,
                            int block_size, int kv_dim, const int* bt_row, const int* pos,
                            int M, cudaStream_t s) {
     int n = M * kv_dim, blk = 256;
